Moved authentication to separate file.
